Rename thread_pool_attached_executors.hpp to match class name
The executor has been renamed to restricted_thread_pool_executor. Still
forwards the old file to the new implementation.
