+++ libs/compute/include/hpx/compute/host/block_allocator.hpp
@@ -20,7 +20,7 @@
 #include <hpx/functional/invoke_fused.hpp>
 #include <hpx/iterator_support/range.hpp>
 #include <hpx/parallel/algorithms/for_each.hpp>
-#include <hpx/execution/executors/thread_pool_attached_executors.hpp>
+#include <hpx/parallel/executors/restricted_thread_pool_executor.hpp>
 #include <hpx/parallel/util/cancellation_token.hpp>
 #include <hpx/parallel/util/partitioner_with_cleanup.hpp>
 #include <hpx/topology/topology.hpp>
