+++ libs/execution/CMakeLists.txt
@@ -94,6 +94,7 @@ set(execution_compat_headers
   hpx/parallel/executors/pool_executor.hpp
   hpx/parallel/executors/post_policy_dispatch.hpp
   hpx/parallel/executors/rebind_executor.hpp
+  hpx/parallel/executors/restricted_thread_pool_executor.hpp
   hpx/parallel/executors/sequenced_executor.hpp
   hpx/parallel/executors/service_executors.hpp
   hpx/parallel/executors/static_chunk_size.hpp
